refactor: use checkboxes fir datalist
